employee, ProductCollection

+++ src/main/java/vn/id/vuductrieu/tlcn_be/entity/mongodb/ProductCollection.java
@@ -35,7 +35,7 @@ public class ProductCollection {
 
     private Integer batterySize;
 
-    private Integer screenSize;
+    private Double screenSize;
 
     private Integer weight;
